Modularize web landing page
Separate route metadata, page composition, hero interactions, privacy visuals, social proof, and pricing without changing rendered behavior.
